+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.0.0
 * Added support for parsing attributes
 * Changed public interface for 'parseContent'
-* Added usefull functions
+* Added useful functions
 * Dropped unnecessary dependency from shakespeare
 
 ## 0.1.0